pyproject.toml
[project]
name = "atliq-hr-assist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.9.4",
    "sqlalchemy>=2.0.0",
    "mysqlclient>=2.1.0",
    "alembic>=1.12.0",
    "python-dotenv>=1.0.0",
]

[tool.poetry.dependencies]
pydantic = "^2.0"
